Update fish status in fish timer window when uptime changes
Currently, if a fish becomes available and you are not currently
fishing, the row continues to be marked red as unavailable.
The time counts down, but it's hard to tell the difference between
"time counting down until the window opens" and
"time counting down until the window closes" when the row color
is the same.