Add far node filter distance to Auto Gather configuration
A new parameter for configuration, 'Far Node Filter Distance', has been added to the Auto Gather Context. This parameter will allow a user to set a distance to filter out any nodes that are closer than this when looking for non-empty nodes. The value is set via a DragFloat in the GUI and is used when moving to a far node.