Add CollectableDebugUi class for auto gathering
A new class called CollectableDebugUi has been added to the AutoGatherUI. This class creates a modal window that displays the last collectability score. The window is only shown when the AddonGatheringMasterpiece isn't visible. This feature aids in debugging the auto gathering feature by providing real-time feedback on the collectability score.