Added timed node precognizance feature to GatherBuddy
This update introduces a new feature to the GatherBuddy automation tool: timed node precognizance. It allows nodes to be perceived for gathering in advance of their actual uptime, customizable up to 600 seconds. The value can be adjusted through the interface configuration tab. The GatherBuddy's time server has also been updated to take into account this precognizance feature when considering the best location for item gathering.