Improve auto-gathering and navigation functionality
This commit optimizes the auto-gathering functionality by checking if the gatherable object matches the job or is a multiple type. It also restructures the execution of auto-gathering and introduces "DoSafetyChecks". The navigation and stuck-checking mechanisms are also updated to provide smoother character movement and fix situations when characters get stuck. A method to stop path-finding if the character reaches a node or when the GP is less than the minimum has been implemented.