Better handling of hidden items and some corner cases
Don't overcap crystals and treasure maps.
Don't use skills if a node doesn't have any item from the gathering list.
Don't use Luck if hidden items are revealed.
Switch to the next item if the node has it and we have enough of the current one.
Don't gather treasure maps if there is no map allowance.
Try not to gather a treasure map if it's not on the list.