Check that actions are unlocked before using them.
Refactoring of action usage checks to reduce the amount of duplicate code.
If an action is unlocked by a quest, check that the quest is completed.
Drop the "II" from Bountiful xxx II in the configuration menu, because Bountiful I would be used if Bountiful II is not unlocked.
Correctly calculate the yield of Bountiful if the second version is not unlocked.