Change job after teleporting
The item gathered first may depend on the distance to nodes, so we delay changing jobs until we know which item is closer.
This could alternatively be done before teleportation by checking the distance to the aetheryte, but this is more complex, and vanilla GatherBuddy changes jobs after teleportation, so I chose this implementation.
Initially, teleportation and job change order were switched to facilitate collectables availability check, which depended on the current job, but this is no longer the case.