First attempt at fixing this stuff.
Adds all the fish I could see from the fish list to the data as patch 7.2. Maybe this is wrong?

Fixes the fish timer window by renaming the fishing spots to include the name of the mission that grants it. Then check by pulling the name of the mission through the "_ToDoList" addon.